Move project headers into include/project/
